# internal/handlers/rice_screenshots.go
package handlers

import (
	"net/http"
	"strings"

	"github.com/ricehub-io/api/internal/errs"
	"github.com/ricehub-io/api/internal/security"
	"github.com/ricehub-io/api/internal/services"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type RiceScreenshotHandler struct {
	svc *services.RiceScreenshotService
}

func NewRiceScreenshotHandler(svc *services.RiceScreenshotService) *RiceScreenshotHandler {
	return &RiceScreenshotHandler{svc}
}

// @Summary Upload one or more screenshots for a rice
// @Tags rices
// @Accept mpfd
// @Produce json
// @Param id path string true "Rice ID (UUID)"
// @Param "files[]" formData file true "Screenshot images (multiple allowed)"
// @Success 201 {object} object "Returns screenshots array"
// @Failure 400 {object} models.ErrorDTO "Missing files"
// @Failure 403 {object} models.ErrorDTO "Forbidden"
// @Security BearerAuth
// @Router /rices/{id}/screenshots [post]
func (h *RiceScreenshotHandler) CreateScreenshot(c *gin.Context) {
	token := c.MustGet("token").(*security.AccessToken)
	userID, _ := uuid.Parse(token.Subject)

	var path ricesPath
	if err := c.ShouldBindUri(&path); err != nil {
		c.Error(errs.InvalidRiceID)
		return
	}
	riceID, _ := uuid.Parse(path.RiceID)

	form, formErr := c.MultipartForm()
	if formErr != nil {
		c.Error(errs.UserError("Invalid multipart form", http.StatusBadRequest))
		return
	}

	files := form.File["files[]"]
	if len(files) == 0 {
		c.Error(errs.MissingFile)
		return
	}

	scrs, err := h.svc.CreateScreenshot(c.Request.Context(), userID, riceID, files, token.IsAdmin)
	if err != nil {
		c.Error(err)
		return
	}

	c.JSON(http.StatusCreated, gin.H{"screenshots": scrs})
}

// @Summary Delete a screenshot from a rice
// @Tags rices
// @Param id path string true "Rice ID (UUID)"
// @Param screenshotId path string true "Screenshot ID (UUID)"
// @Success 204 "Deleted"
// @Failure 400 {object} models.ErrorDTO "Invalid UUID"
// @Failure 403 {object} models.ErrorDTO "Forbidden"
// @Failure 404 {object} models.ErrorDTO "Screenshot not found"
// @Failure 422 {object} models.ErrorDTO "At least one screenshot required"
// @Security BearerAuth
// @Router /rices/{id}/screenshots/{screenshotId} [delete]
func (h *RiceScreenshotHandler) DeleteScreenshot(c *gin.Context) {
	token := c.MustGet("token").(*security.AccessToken)
	userID, _ := uuid.Parse(token.Subject)

	var path struct {
		RiceID       string `uri:"id" binding:"required,uuid"`
		ScreenshotID string `uri:"screenshotId" binding:"required,uuid"`
	}

	if err := c.ShouldBindUri(&path); err != nil {
		msg := err.Error()
		if strings.Contains(msg, "RiceID") {
			msg = errs.InvalidRiceID.Error()
		} else if strings.Contains(msg, "ScreenshotID") {
			msg = "Invalid screenshot ID path parameter. It must be a valid UUID."
		}
		c.Error(errs.UserError(msg, http.StatusBadRequest))
		return
	}
	riceID, _ := uuid.Parse(path.RiceID)
	screenshotID, _ := uuid.Parse(path.ScreenshotID)

	if err := h.svc.DeleteScreenshot(c.Request.Context(), riceID, screenshotID, userID, token.IsAdmin); err != nil {
		c.Error(err)
		return
	}

	c.Status(http.StatusNoContent)
}